assocations + scope specs for collections
